answer_generation/pyproject.toml
[tool.poetry]
name = "answer-generation"
version = "0.1.0"
description = ""
authors = ["MakArtKar <makoyan2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.23"
langchain-core = "^0.2.38"
httpx = "^0.27.2"
ipykernel = "^6.29.5"
langchain = "^0.2.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
